+++ packages/server/src/auth/permissions/study.ts
@@ -1,4 +1,4 @@
-import { Roles } from '../roles';
+import { Roles } from './roles';
 
 /** Permissions associated with studies */
 export enum StudyPermissions {
