+++ packages/server/src/auth/permissions/tag.ts
@@ -1,4 +1,4 @@
-import { Roles } from '../roles';
+import { Roles } from './roles';
 
 export enum TagPermissions {
   CREATE = 'tag:create',
